requirements.txt
flask>=3.0.0
flask-sqlalchemy>=3.1.0
flask-migrate>=4.0.0
flask-redis>=0.4.0
flask-login>=0.6.0
flask-wtf>=1.2.0
marshmallow-sqlalchemy
marshmallow>=3.20.0
celery>=5.3.0
redis>=5.0.0
gunicorn>=21.2.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0
tensorflow>=2.15.0
transformers[tf-cpu]>=4.35.0
huggingface-hub
sentencepiece
vaderSentiment
ftfy
tqdm
pydantic>=2.0
torch>=2.1.0
pandas>=2.1.0
numpy>=1.26.0
alembic>=1.12.0
flask-cors>=4.0.0
